+++ CHANGELOG.rst
@@ -1,5 +1,5 @@
 ..
-	Copyright 2019-2022 MicroEJ Corp. All rights reserved.
+	Copyright 2019-2023 MicroEJ Corp. All rights reserved.
 	Use of this source code is governed by a BSD-style license that can be found with this software.
 
 =========
@@ -10,15 +10,52 @@ All notable changes to this project will be documented in this file.
 
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_, and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
-2.8.0 -- 2022-07-23
--------------------
+------------------
+2.9.0 - 2023-06-29
+------------------
+
+Changed
+=======
+
+- Platform Configuration Additions 2.0.0 (see `CHANGELOG <framework/platform/content-sdk-5/build/CHANGELOG.md>`_).
+
+
+Added 
+=====
+
+- MicroEJ External Resource Loader Validation (Java) 1.0.0 (see `CHANGELOG <tests/ext-res-loader/java-testsuite-runner-ext-res-loader/CHANGELOG.rst>`_).
+- MicroEJ MicroVG Validation (Java) 1.0.0 (see `CHANGELOG <tests/microvg/java-testsuite-runner-microvg/CHANGELOG.rst>`_).
+
+Changed
+=======
+
+- MicroEJ UI Validation (Java) 1.2.0 (see `CHANGELOG <tests/ui/ui3/java-testsuite-runner-ui3/CHANGELOG.rst>`_).
+- MicroEJ Core Validation (Java) 3.2.0 (see `CHANGELOG <tests/core/java-testsuite-runner-core/CHANGELOG.rst>`_).
+- MicroEJ ECOM-WIFI Validation (Java) 1.0.1 (see `CHANGELOG <tests/ecom-wifi/java-testsuite-runner-ecom-wifi/CHANGELOG.rst>`_).
+- MicroEJ FS Validation (Java) 1.0.1 (see `CHANGELOG <tests/fs/java-testsuite-runner-fs/CHANGELOG.rst>`_).
+- MicroEJ NET Validation (Java) 1.0.1 (see `CHANGELOG <tests/net/java-testsuite-runner-net/CHANGELOG.rst>`_).
+- MicroEJ SECURITY Validation (Java) 1.0.1 (see `CHANGELOG <tests/security/java-testsuite-runner-security/CHANGELOG.rst>`_).
+- MicroEJ SSL Validation (Java) 1.0.1 (see `CHANGELOG <tests/ssl/java-testsuite-runner-ssl/CHANGELOG.rst>`_).
+- MicroEJ WATCHDOG-TIMER Validation (Java) 1.0.1 (see `CHANGELOG <tests/watchdog-timer/java-testsuite-runner-watchdog-timer/CHANGELOG.rst>`_).
+
+Fixed
+=====
+
+- Fix broken links. 
+- Fix option name in ``config.properties.tpl`` files (No BSP Connection: use ``deploy.dir.microejscript`` instead of ``deploy.bsp.microejscript``). 
+
+------------------
+2.8.0 - 2022-07-23
+------------------
 
 Added
 =====
 
-- Add NET Test Suite for MicroEJ NET Packs 10.0.0 and higher (see `README <tests/net/README.rst>`_).
-- Add Watchdog Timer Test Suite for MicroEJ Watchdog Timer Packs 2.0.1 and higher (see `README <tests/watchdog-timer/README.rst>`_).
-- Add ECOM-WIFI Test Suite for ECOM-WIFI 2.1 and higher (see `README <tests/ecom-wifi/README.rst>`_).
+- MicroEJ NET Validation (Java) 1.0.0 for MicroEJ NET Packs 10.0.0 and higher (see `README <tests/net/README.rst>`_).
+- MicroEJ SSL Validation (Java) 1.0.0 for MicroEJ NET Packs 10.0.0 and higher (see `README <tests/ssl/README.rst>`_).
+- MicroEJ SECURITY Validation (Java) 1.0.0 for MicroEJ NET Packs 10.0.0 and higher (see `README <tests/security/README.rst>`_).
+- MicroEJ WATCHDOG-TIMER Validation (Java) 1.0.0 for MicroEJ WATCHDOG-TIMER Packs 2.0.1 and higher (see `README <tests/watchdog-timer/README.rst>`_).
+- MicroEJ ECOM-WIFI Validation (Java) 1.0.0 for MicroEJ ECOM-WIFI 2.1 and higher (see `README <tests/ecom-wifi/README.rst>`_).
 
 Changed
 =======
@@ -32,8 +69,9 @@ Remove
 
 - MicroEJ UI Validation: remove the C project.
 
-2.7.0 -- 2021-07-16
--------------------
+------------------
+2.7.0 - 2021-07-16
+------------------
 
 Changed
 =======
@@ -45,8 +83,9 @@ Fixed
 
 - Fixed links and paths in Graphical User Interface Test Suite README files and Core Engine Test Suite README file.
 
-2.6.0 -- 2021-06-25
--------------------
+------------------
+2.6.0 - 2021-06-25
+------------------
 
 Added
 =====
@@ -58,8 +97,9 @@ Changed
 
 - Graphical User Interface Test Suite for MicroEJ UI Packs [6.0.0-13.0.0[ (13.0.0 excluded) has moved from ``tests/ui/`` to ``tests/ui/ui2/`` (see `README <tests/ui/ui2/README.rst>`_).
 
-2.5.0 -- 2021-05-11
--------------------
+------------------
+2.5.0 - 2021-05-11
+------------------
 
 Changed
 =======
@@ -73,9 +113,9 @@ Fixed
 
 - URLs in the documentation.
 
--------------------
-2.4.0 -- 2021-04-16
--------------------
+------------------
+2.4.0 - 2021-04-16
+------------------
 
 Added
 =====
@@ -99,9 +139,9 @@ Fixed
 - Documentation of ``target.platform.dir`` to setup the Test Suite FS.
 - CORE Test Suite updated to 3.0.2 (see `CHANGELOG <tests/core/java-testsuite-runner-core/CHANGELOG.rst`_).
 
--------------------
-2.3.0 -- 2021-01-13
--------------------
+------------------
+2.3.0 - 2021-01-13
+------------------
 
 The main changes of this release are the addition of the Java Test
 Suite Filesystem for FS API and the Platform Configuration Additions